update stoichmetric precipitator
